pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "game-studio-ai"
version = "0.1.0"
description = "Tiered multi-agent game development orchestrator"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.30.0",
    "pydantic>=2.0",
    "typer>=0.12.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlmodel>=0.0.18",
    "pyyaml>=6.0",
    "Pillow>=10.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
]

[project.scripts]
studio = "runner:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]
